Renamed `import` package to `result`
