+++ internal/models/result/execution.go
@@ -1,4 +1,4 @@
-package _import
+package result
 
 import "time"
