+++ internal/models/result/step.go
@@ -1,4 +1,4 @@
-package _import
+package result
 
 type Step struct {
 	Data      Data
