+++ internal/models/result/suite.go
@@ -1,4 +1,4 @@
-package _import
+package result
 
 type Suite struct {
 	Data []SuiteData
